ecosync_project/requirements.txt
Django
drf-nested-routers
djangorestframework
reportlab
psycopg2
requests